refactor to separate application logic from default and testing server running modes
