+++ requirements.txt
@@ -3,3 +3,4 @@ requests
 SQLAlchemy
 tornado
 bleach
+pytest
