--- src/aswwu/default.conf
@@ -1,2 +1,0 @@
-log_name = "aswwu"
-current_year = "1920"
