+++ src/aswwu/route_handlers/elections.py
@@ -7,7 +7,7 @@ from src.aswwu.base_handlers import BaseHandler
 import src.aswwu.exceptions as exceptions
 from src.aswwu.permissions import permission_and, admin_permission, elections_permission
 
-from src import aswwu as elections_alchemy
+import src.aswwu.alchemy_new.elections as elections_alchemy
 import src.aswwu.alchemy_new.mask as mask_alchemy
 import src.aswwu.models.elections as elections_model
 import src.aswwu.models.mask as mask_model
