+++ tests/default.conf
@@ -1,0 +1,3 @@
+port=8888
+log_file="aswwu"
+current_year="1920"
